Updated this file with two Mega Colors.
MegaRed and MegaBlue are what you think they are: colors. These two colors are designed with a key difference that separates them from the rest of the colors: They are pure red and blue.

When I coded in MegaRed and MegaBlue, I assumed it was a RGBA scale. I just hope it's not actually ARGB, otherwise it may not work. :P